add editor and crud img

+++ src/pages/api/dashboard/service/[id].ts
@@ -1,6 +1,7 @@
 import { getSession } from "next-auth/react";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { prisma } from "@/lib/prismadb";
+import fs from "fs";
 
 export default async function handler(
   req: NextApiRequest,
@@ -13,17 +14,23 @@ export default async function handler(
   }
   if (req.method === "DELETE") {
     const { id } = req.query;
+    const page = await prisma.pages.findUnique({
+      where: { id: id as string },
+    });
+    if (!page) {
+      return res.status(404).json({ message: "Page not found." });
+    }
+    const { cover } = page;
+    // Delete image from public/uploads folder
+    fs.unlinkSync(`public/${cover}`);
     const deletePage = await prisma.pages.delete({
-      where: {
-        id: id as string,
-      },
+      where: { id: id as string },
     });
     res.status(200).json(deletePage);
   }
-
   // HTTP method not supported!
   else {
-    res.setHeader("Allow", ["GET"]);
+    res.setHeader("Allow", ["DELETE"]);
     res
       .status(405)
       .json({ message: `HTTP method ${req.method} is not supported.` });
